Tower: change the way it is instantiated and how it instantiate the Floors
Also heavily changes the towers.json is structured: it's loaded by Board
and contains all the information about the required towers and their
floors (i.e. it lists which towers contains which type of cards, and
how many floors it has and which resources these floors will give out as
a bonus.

Signed-off-by: Mattia Rizzolo <mattia@mapreri.org>